Combine label index and inherit index into single type.
Allows for them to share cached data, reducing occupancy.
