+++ labelindex/label_inherit_index_test.go
@@ -22,7 +22,6 @@ import (
 
 	"github.com/projectcalico/libcalico-go/lib/backend/model"
 	"github.com/projectcalico/libcalico-go/lib/selector"
-	"github.com/projectcalico/libcalico-go/lib/selector/parser"
 )
 
 type update struct {
@@ -46,7 +45,7 @@ var _ = Describe("Keys", func() {
 var _ = Describe("Index", func() {
 	var (
 		updates []update
-		idx     Index
+		idx     *InheritIndex
 		a_eq_a1 selector.Selector
 		a_eq_b  selector.Selector
 		c_eq_d  selector.Selector
@@ -68,7 +67,7 @@ var _ = Describe("Index", func() {
 
 	BeforeEach(func() {
 		updates = make([]update, 0)
-		idx = NewIndex(onMatchStart, onMatchStop)
+		idx = NewInheritIndex(onMatchStart, onMatchStop)
 
 		a_eq_a1, err = selector.Parse(`a=="a1"`)
 		Expect(err).To(BeNil())
@@ -80,8 +79,8 @@ var _ = Describe("Index", func() {
 
 	Context("with empty index", func() {
 		It("should do nothing when adding labels", func() {
-			idx.UpdateLabels("foo", parser.MapAsLabels(map[string]string{"a": "b"}))
-			idx.UpdateLabels("bar", parser.MapAsLabels(map[string]string{}))
+			idx.UpdateLabels("foo", map[string]string{"a": "b"}, nil)
+			idx.UpdateLabels("bar", map[string]string{}, nil)
 			Expect(updates).To(BeEmpty())
 		})
 		It("should do nothing when adding selectors", func() {
@@ -93,8 +92,7 @@ var _ = Describe("Index", func() {
 
 	Context("with one set of labels added", func() {
 		BeforeEach(func() {
-			idx.UpdateLabels("l1",
-				parser.MapAsLabels(map[string]string{"a": "b", "c": "d"}))
+			idx.UpdateLabels("l1", map[string]string{"a": "b", "c": "d"}, nil)
 		})
 
 		It("should ignore non-matching selectors", func() {
@@ -108,7 +106,7 @@ var _ = Describe("Index", func() {
 		It("should fire correct events for matching selector", func() {
 			By("firing start event on addition")
 			idx.UpdateSelector("e1", a_eq_b)
-			Expect(updates).To(Equal([]update{update{
+			Expect(updates).To(Equal([]update{{
 				"start", "l1", "e1",
 			}}))
 			updates = updates[:0]
@@ -120,7 +118,7 @@ var _ = Describe("Index", func() {
 			Expect(updates).To(BeEmpty())
 			By("firing stop event on deletion")
 			idx.DeleteSelector("e1")
-			Expect(updates).To(Equal([]update{update{
+			Expect(updates).To(Equal([]update{{
 				"stop", "l1", "e1",
 			}}))
 		})
@@ -130,22 +128,22 @@ var _ = Describe("Index", func() {
 			idx.UpdateSelector("e1", a_eq_b)
 			idx.UpdateSelector("e2", c_eq_d)
 			Expect(updates).To(Equal([]update{
-				update{"start", "l1", "e1"},
-				update{"start", "l1", "e2"},
+				{"start", "l1", "e1"},
+				{"start", "l1", "e2"},
 			}))
 			updates = updates[:0]
 
 			By("firing stop for update to non-matching selector")
 			idx.UpdateSelector("e2", a_eq_a1)
 			Expect(updates).To(Equal([]update{
-				update{"stop", "l1", "e2"},
+				{"stop", "l1", "e2"},
 			}))
 			updates = updates[:0]
 
 			By("firing stop when selector deleted")
 			idx.DeleteSelector("e1")
 			Expect(updates).To(Equal([]update{
-				update{"stop", "l1", "e1"},
+				{"stop", "l1", "e1"},
 			}))
 		})
 	})
@@ -156,51 +154,51 @@ var _ = Describe("Index", func() {
 		})
 
 		It("should ignore non-matching labels", func() {
-			idx.UpdateLabels("l1", parser.MapAsLabels(map[string]string{"a": "b"}))
+			idx.UpdateLabels("l1", map[string]string{"a": "b"}, nil)
 			Expect(updates).To(BeEmpty())
 		})
 		It("should fire correct events for match", func() {
 			By("firing for add")
-			idx.UpdateLabels("l1", parser.MapAsLabels(map[string]string{"a": "a1"}))
-			Expect(updates).To(Equal([]update{update{
+			idx.UpdateLabels("l1", map[string]string{"a": "a1"}, nil)
+			Expect(updates).To(Equal([]update{{
 				"start", "l1", "e1",
 			}}))
 			updates = updates[:0]
 			By("ignoring idempotent add")
-			idx.UpdateLabels("l1", parser.MapAsLabels(map[string]string{"a": "a1"}))
+			idx.UpdateLabels("l1", map[string]string{"a": "a1"}, nil)
 			Expect(updates).To(BeEmpty())
 			By("ignoring update to also-matching labels")
 			idx.UpdateLabels("l1",
-				parser.MapAsLabels(map[string]string{"a": "a1", "b": "c"}))
+				map[string]string{"a": "a1", "b": "c"}, nil)
 			Expect(updates).To(BeEmpty())
 			By("firing stop on delete")
 			idx.DeleteLabels("l1")
-			Expect(updates).To(Equal([]update{update{
+			Expect(updates).To(Equal([]update{{
 				"stop", "l1", "e1",
 			}}))
 		})
 		It("should handle multiple matches", func() {
 			By("firing events for both")
-			idx.UpdateLabels("l1", parser.MapAsLabels(map[string]string{"a": "a1"}))
+			idx.UpdateLabels("l1", map[string]string{"a": "a1"}, nil)
 			idx.UpdateLabels("l2",
-				parser.MapAsLabels(map[string]string{"a": "a1", "b": "b1"}))
+				map[string]string{"a": "a1", "b": "b1"}, nil)
 			Expect(updates).To(Equal([]update{
-				update{"start", "l1", "e1"},
-				update{"start", "l2", "e1"},
+				{"start", "l1", "e1"},
+				{"start", "l2", "e1"},
 			}))
 			updates = updates[:0]
 
 			By("handling updates to non-matching labels")
-			idx.UpdateLabels("l1", parser.MapAsLabels(map[string]string{"a": "a2"}))
+			idx.UpdateLabels("l1", map[string]string{"a": "a2"}, nil)
 			Expect(updates).To(Equal([]update{
-				update{"stop", "l1", "e1"},
+				{"stop", "l1", "e1"},
 			}))
 			updates = updates[:0]
 
 			By("handling removal of selector")
 			idx.DeleteSelector("e1")
 			Expect(updates).To(Equal([]update{
-				update{"stop", "l2", "e1"},
+				{"stop", "l2", "e1"},
 			}))
 		})
 	})
